requirements.txt
mysql_connector_repackaged==0.3.1
openpyxl==3.0.10
paho_mqtt==1.6.1
PyQt5==5.15.9
PyQt5_sip==12.11.0
xlwings==0.30.10
